pyproject.toml
[tool.poetry]
name = "test-for-jetlend"
version = "0.1.0"
description = "tasks for test"
authors = ["Stanis96 <vasilievstas96@yandex.ru>"]
readme = "README.md"
packages = [{include = "test_for_jetlend"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
bs4 = "^0.0.1"
beautifulsoup4 = "^4.11.1"
psycopg2-binary = "^2.9.5"
pandas = "^1.5.1"
openpyxl = "^3.0.10"
django = "^4.1.3"
django-rest-framework = "^0.1.0"
celery = "^5.2.7"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
pyproject-flake8 = "^0.0.1-alpha.4"
flake8-black = "^0.3.3"
flake8-builtins = "^1.5.3"
flake8-import-order = "^0.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
exclude = """
    ('
    |.idea/
    |.venv/
    |.vscode/
    ')
"""

# call using pflake8
[tool.flake8]
max-line-length = 100
max-complexity = 8
ignore = "A003, W503"
exclude = """
    .idea/
    .venv/
    .vscode/
    */migrations/
"""

[virtualenvs]
create = true
in-project = true
path = ".venv"
